src/requirements.txt
google-adk==0.4.0
python-dotenv==1.1.0
pandas==2.2.3
tqdm